# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API Documentation",
    "version": "1.0.0",
    "description": "API documentation for my Next.js app"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://story-teller-batrinutsilvius-projects.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token from the login endpoint"
      }
    }
  },
  "paths": {
...
    "/api/stories/{story_id}/categories/language/{language_id}": {
      "get": {
        "summary": "Gets all the categories of a story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The story ID"
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The language ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories for the story with translated names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "StoryCategory ID"
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "categoryTranslations": {
                            "type": "array",
                            "description": "Contains one translation for the requested language",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/stories/{story_id}/languages/{language_id}/minified": {
      "get": {
        "summary": "Gets a summary story translation by language",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Story summary without pages (title, description, cover and duration)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "StoryTranslation ID"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "story": {
                        "type": "object",
                        "properties": {
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "duration": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Story duration in seconds"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/stories/{story_id}/languages/{language_id}": {
      "get": {
        "summary": "Gets a story translation by language",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pages",
            "schema": {
              "type": "integer",
              "default": 5
            },
            "description": "Number of pages to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Story translation with paginated pages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "StoryTranslation ID"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "story": {
                        "type": "object",
                        "properties": {
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      },
                      "language": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "country_code": {
                            "type": "string"
                          }
                        }
                      },
                      "storyPages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "page_number": {
                              "type": "integer"
                            },
                            "text_content": {
                              "type": "string"
                            },
                            "photo_url": {
                              "type": "string",
                              "nullable": true
                            },
                            "audio_url": {
                              "type": "string",
                              "nullable": true
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}